requirements.txt
discord.py==2.4.0
aiohttp==3.11.11
PyYAML==6.0.1
APScheduler==3.10.4
pytz==2024.1
